refactoring: rename component
